pour eviter une modification d'un user qui ne serait pas auteur de post en acces url

+++ app/Http/Controllers/PostsController.php
@@ -76,6 +76,9 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Non autorisé');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -111,6 +114,9 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Non autorisé');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Votre article a été supprimé ! ');
     }
